MPFR 3.1.2 tag.

+++ tests/texp.c
@@ -20,6 +20,10 @@ along with the GNU MPFR Library; see the file COPYING.LESSER.  If not, see
 http://www.gnu.org/licenses/ or write to the Free Software Foundation, Inc.,
 51 Franklin St, Fifth Floor, Boston, MA 02110-1301, USA. */
 
+#include <stdio.h>
+#include <stdlib.h>
+#include <limits.h>
+
 #include "mpfr-test.h"
 
 #ifdef CHECK_EXTERNAL
@@ -712,8 +716,7 @@ underflow_up (int extended_emin)
                         printf (" and extended emin");
                       printf ("\nfor precx = %d, precy = %d, %s\n",
                               precx, precy, e3 ? "mpfr_exp_3" : "mpfr_exp");
-                      printf ("Got %u instead of %u.\n",
-                              (unsigned int) __gmpfr_flags,
+                      printf ("Got %u instead of %u.\n", __gmpfr_flags,
                               (unsigned int) MPFR_FLAGS_INEXACT);
                       err = 1;
                     }
@@ -837,7 +840,7 @@ underflow_up (int extended_emin)
                                   precy + i, 1 - 2 * (precy + i));
                         printf (", %s\n", e3 ? "mpfr_exp_3" : "mpfr_exp");
                         printf ("Got %u instead of %u.\n",
-                                (unsigned int) __gmpfr_flags, flags);
+                                __gmpfr_flags, flags);
                         err = 1;
                       }
                     if (rnd == MPFR_RNDU || rnd == MPFR_RNDA || rnd == MPFR_RNDN ?
