+++ tests/tpow_z.c
@@ -20,6 +20,7 @@ along with the GNU MPFR Library; see the file COPYING.LESSER.  If not, see
 http://www.gnu.org/licenses/ or write to the Free Software Foundation, Inc.,
 51 Franklin St, Fifth Floor, Boston, MA 02110-1301, USA. */
 
+#include <stdlib.h>
 #include <float.h>
 #include <math.h>
 
@@ -249,8 +250,7 @@ bug20071104 (void)
     }
   if (__gmpfr_flags != (MPFR_FLAGS_OVERFLOW | MPFR_FLAGS_INEXACT))
     {
-      printf ("Error in bug20071104: bad flags (%u)\n",
-              (unsigned int) __gmpfr_flags);
+      printf ("Error in bug20071104: bad flags (%u)\n", __gmpfr_flags);
       exit (1);
     }
   mpfr_clears (x, y, (mpfr_ptr) 0);
